+++ src/app/profile/page.tsx
@@ -1,4 +1,4 @@
-import ProfilePageClient from '@/components/profile/ProfilePage'
+import ProfilePageClient from '@/components/profile/profile-page'
 
 export default function ProfilePage() {
   return <ProfilePageClient />
