+++ src/app/profile/profile-edit/page.tsx
@@ -1,4 +1,4 @@
-import ProfileEditClient from '@/components/profile-edit/ProfileEditPage'
+import ProfileEditClient from '@/components/profile-edit/profile-edit-page'
 
 export default function ProfileEditPage() {
   return <ProfileEditClient />
